Add client-side proxies for Earth Engine objects
In order to run ordinations client-side, there needs to be a mechanism
for translating `ee.FeatureCollections` to python objects (in this case,
pydantic models).  Based on work from @aazuspan, these classes provide
multi-threaded functionality to convert feature collections in a
lower-memory context.  `Feature` and `FeatureCollection` have a few
helper functions to extract properties as values and lists.