WIP: First cut at integrating sknnr estimators through client-side
This commit implements client-side fitting of four of the five
estimators (excluding Raw) and implements them using sknnr transformers
associated with each method.  For this commit only, we have introduced
a new method in each estimator called `train_client` whilst retaining
the original `train` which is a server-side train/fit.  The general
workflow is to convert the training data into a client-side feature
collection (class FeatureCollection), fit the model using the
transformer, and convert all needed data objects back into server-side
objects to work with `predict` and `predict_fc`.  For all but GNN,
`predict` and `predict_fc` are unchanged and tests are passing for both
client-side and server-side training.  For GNN, there were slight
modifications to `predict` and `predict_fc` in the transformations such
that only client-side fitting is working.  There are expected test
failures in this commit.

In addition, we have dropped support for python 3.8 in this commit as
types `list` and `dict` were not compatible with pydantic
BaseModel-derived classes for 3.8.

Still, much refactoring needs to be done to remove duplication across
the methods.