Large refactor of most estimators into `Transformed` class
This commit changes the class hierachy to reflect that of `sknnr`.
The `GeeKnnClassifier` class has been renamed to `Raw`, a new abstract
class called `Transformed` derives from `Raw` and all other estimators
(`Eucliean`, `Mahalanobis`, `MSN`, and `GNN`) are all subclasses of
`Transformed`.  `Transformed` does most of the work in training and
predicting the estimators and four abstract methods delegate the
responsiblity of defining the transformer to the subclasses.

Additionally, `cca.py` and `ccora.py` are no longer needed as all
model fitting is now done through `sknnr`.