Create client-side ordinations to avoid server-memory issues (#4)
This commit integrates `scikit-learn-knn-regression` (aka `sknnr`)
estimators as the basis for client-side fitting (training) of 
gee-knn estimators.  Previously, we had implemented all estimators
using the GEE API which meant that training and predicting was all
running server-side.  With this commit, the general workflow is to
convert the training data into a client-side feature collection
(class `FeatureCollection`), fit the model using the `sknnr`
transformers, and convert all needed data objects back into server-side
objects to work with `predict`.  We now have a single source library
for model fitting (`sknnr`), but now two backends for spatial
predictions - `sknnr-spatial` for local predictions and `geeknn` for
GEE-based spatial predictions.  

This commit also featured heavy refactoring of the estimator classes
and substantial changes to the API.  It drops support for Python 3.8
and adds support for Python 3.12.